Update dependencies to support MongoDB 3.4.x
In order for Monger to work with MongoDB 3.4.x and higher, the Java driver must also be 3.4.x and higher.  Additionally, the 3.4.x driver series will generally work with Java version 6.x, it requires 7.x, and even 8.x for specific SSL functions that are needed by some "Platform As A Service" providers such as MongoDB Atlas.  Here are some relevant links that provide some additional details: 

- http://mongodb.github.io/mongo-java-driver/3.4/upgrading/
- https://docs.atlas.mongodb.com/driver-connection/#java-driver-example
- http://mongodb.github.io/mongo-java-driver/?_ga=2.85301265.905224850.1494354505-76319364.1488584191